+++ plugins/tapd/tasks/bug_converter.go
@@ -82,7 +82,7 @@ func ConvertBug(taskCtx core.SubTaskContext) errors.Error {
 				OriginalStatus: toolL.Status,
 			}
 			if domainL.ResolutionDate != nil && domainL.CreatedDate != nil {
-				domainL.LeadTimeMinutes = uint(domainL.ResolutionDate.Sub(*domainL.CreatedDate).Minutes())
+				domainL.LeadTimeMinutes = int64(domainL.ResolutionDate.Sub(*domainL.CreatedDate).Minutes())
 			}
 			results := make([]interface{}, 0, 2)
 			boardIssue := &ticket.BoardIssue{
